@@ -14,13 +14,13 @@ description: A new Flutter application.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: ">=2.15.1 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  bonfire: ^2.0.0
+  bonfire: ^2.1.0
   socket_io_client: ^1.0.2
 
 dev_dependencies:
